+++ pkg/tsdb/graphite/graphite.go
@@ -2,14 +2,20 @@ package graphite
 
 import (
 	"encoding/json"
+	"fmt"
 	"io/ioutil"
 	"net/http"
 	"net/url"
+	"path"
 	"strings"
 	"time"
 
 	"github.com/grafana/grafana/pkg/log"
 	"github.com/grafana/grafana/pkg/tsdb"
+)
+
+var (
+	HttpClient = http.Client{Timeout: time.Duration(10 * time.Second)}
 )
 
 type GraphiteExecutor struct {
@@ -30,7 +36,7 @@ func init() {
 func (e *GraphiteExecutor) Execute(queries tsdb.QuerySlice, context *tsdb.QueryContext) *tsdb.BatchResult {
 	result := &tsdb.BatchResult{}
 
-	params := url.Values{
+	formData := url.Values{
 		"from":          []string{"-" + formatTimeRange(context.TimeRange.From)},
 		"until":         []string{formatTimeRange(context.TimeRange.To)},
 		"format":        []string{"json"},
@@ -38,28 +44,24 @@ func (e *GraphiteExecutor) Execute(queries tsdb.QuerySlice, context *tsdb.QueryC
 	}
 
 	for _, query := range queries {
-		params["target"] = []string{query.Query}
-		glog.Debug("Graphite request", "query", query.Query)
+		formData["target"] = []string{query.Query}
 	}
 
-	client := http.Client{Timeout: time.Duration(10 * time.Second)}
-	res, err := client.PostForm(e.Url+"/render?", params)
+	glog.Info("Graphite request body", "formdata", formData.Encode())
+
+	req, err := e.createRequest(formData)
 	if err != nil {
 		result.Error = err
 		return result
 	}
-	defer res.Body.Close()
-
-	body, err := ioutil.ReadAll(res.Body)
+	res, err := HttpClient.Do(req)
 	if err != nil {
 		result.Error = err
 		return result
 	}
 
-	var data []TargetResponseDTO
-	err = json.Unmarshal(body, &data)
+	data, err := e.parseResponse(res)
 	if err != nil {
-		glog.Info("Failed to unmarshal graphite response", "error", err, "body", string(body))
 		result.Error = err
 		return result
 	}
@@ -75,6 +77,46 @@ func (e *GraphiteExecutor) Execute(queries tsdb.QuerySlice, context *tsdb.QueryC
 
 	result.QueryResults["A"] = queryRes
 	return result
+}
+
+func (e *GraphiteExecutor) parseResponse(res *http.Response) ([]TargetResponseDTO, error) {
+	body, err := ioutil.ReadAll(res.Body)
+	defer res.Body.Close()
+	if err != nil {
+		return nil, err
+	}
+
+	if res.StatusCode == http.StatusUnauthorized {
+		glog.Info("Request is Unauthorized", "status", res.Status, "body", string(body))
+		return nil, fmt.Errorf("Request is Unauthorized status: %v body: %s", res.Status, string(body))
+	}
+
+	var data []TargetResponseDTO
+	err = json.Unmarshal(body, &data)
+	if err != nil {
+		glog.Info("Failed to unmarshal graphite response", "error", err, "status", res.Status, "body", string(body))
+		return nil, err
+	}
+
+	return data, nil
+}
+
+func (e *GraphiteExecutor) createRequest(data url.Values) (*http.Request, error) {
+	u, _ := url.Parse(e.Url)
+	u.Path = path.Join(u.Path, "render")
+
+	req, err := http.NewRequest(http.MethodPost, u.String(), strings.NewReader(data.Encode()))
+	if err != nil {
+		glog.Info("Failed to create request", "error", err)
+		return nil, fmt.Errorf("Failed to create request. error: %v", err)
+	}
+
+	req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
+	if e.BasicAuth {
+		req.SetBasicAuth(e.BasicAuthUser, e.BasicAuthPassword)
+	}
+
+	return req, err
 }
 
 func formatTimeRange(input string) string {